// src/generated-code/angular/horseless-odata-content-workspace/horselesscontent/microsoft/graph/user.entity.ts
export interface User extends DirectoryObject {
  //#region ODataApiGen Properties
  accountEnabled?: boolean;
  ageGroup?: string;
  assignedLicenses: AssignedLicense[];
  assignedPlans: AssignedPlan[];
  businessPhones: string[];
  city?: string;
  companyName?: string;
  consentProvidedForMinor?: string;
  country?: string;
  creationType?: string;
  department?: string;
  displayName?: string;
  employeeId?: string;
  externalUserState?: string;
  externalUserStateChangeDateTime?: Date;
  faxNumber?: string;
  givenName?: string;
  identities?: ObjectIdentity[];
  imAddresses?: string[];
  isResourceAccount?: boolean;
  jobTitle?: string;
  lastPasswordChangeDateTime?: Date;
  legalAgeGroupClassification?: string;
  licenseAssignmentStates?: LicenseAssignmentState[];
  mail?: string;
  mailNickname?: string;
  mobilePhone?: string;
  onPremisesDistinguishedName?: string;
  onPremisesExtensionAttributes?: OnPremisesExtensionAttributes;
  onPremisesImmutableId?: string;
  onPremisesLastSyncDateTime?: Date;
  onPremisesProvisioningErrors?: OnPremisesProvisioningError[];
  onPremisesSecurityIdentifier?: string;
  onPremisesSyncEnabled?: boolean;
  onPremisesDomainName?: string;
  onPremisesSamAccountName?: string;
  onPremisesUserPrincipalName?: string;
  otherMails: string[];
  passwordPolicies?: string;
  passwordProfile?: PasswordProfile;
  officeLocation?: string;
  postalCode?: string;
  preferredLanguage?: string;
  provisionedPlans: ProvisionedPlan[];
  proxyAddresses: string[];
  showInAddressList?: boolean;
  signInSessionsValidFromDateTime?: Date;
  state?: string;
  streetAddress?: string;
  surname?: string;
  usageLocation?: string;
  userPrincipalName?: string;
  userType?: string;
  mailboxSettings?: MailboxSettings;
  deviceEnrollmentLimit: number;
  aboutMe?: string;
  birthday: Date;
  hireDate: Date;
  interests?: string[];
  mySite?: string;
  pastProjects?: string[];
  preferredName?: string;
  responsibilities?: string[];
  schools?: string[];
  skills?: string[];
  appRoleAssignments?: AppRoleAssignment[];
  ownedDevices?: DirectoryObject[];
  registeredDevices?: DirectoryObject[];
  manager?: DirectoryObject;
  directReports?: DirectoryObject[];
  memberOf?: DirectoryObject[];
  createdObjects?: DirectoryObject[];
  oauth2PermissionGrants?: OAuth2PermissionGrant[];
  ownedObjects?: DirectoryObject[];
  licenseDetails?: LicenseDetails[];
  transitiveMemberOf?: DirectoryObject[];
  outlook?: OutlookUser;
  messages?: Message[];
  mailFolders?: MailFolder[];
  calendar?: Calendar;
  calendars?: Calendar[];
  calendarGroups?: CalendarGroup[];
  calendarView?: Event[];
  events?: Event[];
  people?: Person[];
  contacts?: Contact[];
  contactFolders?: ContactFolder[];
  inferenceClassification?: InferenceClassification;
  photo?: ProfilePhoto;
  photos?: ProfilePhoto[];
  drive?: Drive;
  drives?: Drive[];
  followedSites?: Site[];
  extensions?: Extension[];
  managedDevices?: ManagedDevice[];
  managedAppRegistrations?: ManagedAppRegistration[];
  deviceManagementTroubleshootingEvents?: DeviceManagementTroubleshootingEvent[];
  planner?: PlannerUser;
  insights?: OfficeGraphInsights;
  settings?: UserSettings;
  onenote?: Onenote;
  activities?: UserActivity[];
  onlineMeetings?: OnlineMeeting[];
  joinedTeams?: Team[];
  //#endregion
}